Demo 02/requirements.txt
wordcloud
numpy
six
python-igraph
pandas
matplotlib
pillow
